- Implementando ao interceptador, captura de Erros(msg) e tambem captura de codigos 200 e 204 de sucesso. - Implementando dialogo de confirmacao antes de deletar usuario.

+++ src/app/components/usuario/usuario.component.ts
@@ -32,11 +32,13 @@ export class UsuarioComponent implements OnInit {
   }
 
   public deleteUsuario(id: Number) {
-    this.usuarioService.deleteUserByID(id).subscribe(() => {
-      this.usuarioService.findAllUsers().subscribe(atualizarLista => {
-        this.usuarios = atualizarLista;
+    if (confirm(`Voce confirma que deseja deletar esse usuario?`)) {
+      this.usuarioService.deleteUserByID(id).subscribe(() => {
+        this.usuarioService.findAllUsers().subscribe(atualizarLista => {
+          this.usuarios = atualizarLista;
+        })
       })
-    })
+    }
   }
 
   public findUserByName($event: KeyboardEvent) {
